Replace the specific parsing code with generic code that calls the parsers
Parsing code is now split out to separate parsing objects, so the specific
code for parsing the Garmin code into NMEA and then parsing that is no longer
needed. Likewise the code that gathers NMEA sentences can be removed.

Instead we now detect what type of device it is (Garmin or NMEA), create a
parser for that type and pass any received data to it.