Fix a memory corruption issue when parsing NMEA sentences
We assumed that if there was a \r in the buffer, then there must also be a \n.
if the \r was the final character in the buffer there there would not be a \n
so adding 2 to the sentence length would cause an invalid read off the end of
the buffer, and by making chars_in_buffer equal -1 an invalid write by
underflowing the buffer, scribbling over the address of ctxt. Which then caused
a crash.