Remove an additional memcpy when reading from the socket.
Each time we read into a temporary buffer and the memcpy'd the data into the
real buffer. Instead of using the temporary buffer we now ask the parser module
to give us a pointer to the buffer to write to and how much space there is in it
which means we don't need to memcpy.