Only open TTY devices once
TTY devices were accidentally being opened twice but only closed once. This
meant that when a device file was removed, but then replaced, a file descriptor for the old device was cached somewhere in the deep bowels of the system and
this dead file descriptor was being returned when Gypsy attempted to open the
new device, causing I/O errors when trying to read from it. Which in turn
caused some lovely CPU pegging as Gypsy would just enter a busy loop constantly
trying to read from the dead descriptor.

In short, everyone's a winner!